+++ CHANGELOG.md
@@ -5,9 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.5] - 2024-10-07
+
+### Added
+
+- Support for ghost layers of more than a single element in thickness.
+
+## [0.5.4] - 2024-08-28
+
+### Added
+
+- Function `array_of_tuples`.
+- Export statement for `local_permutation`.
+- Experimental support for block arrays via types `BRange`, `BVector`, `BMatrix`, and `BArray`.
+
 ## [0.5.3] - 2024-08-16
 
 ### Fixed
+
 - Typo: `node_coorinates_unit_cube` -> `node_coordinates_unit_cube`.
 - Bug in `nullspace_linear_elasticity`.
 - Bug in `PVector` when working in split format.